requirements.txt
feature_engine==1.9.3
kaggle==2.0.0
matplotlib==3.10.8
mlflow==3.10.0
pandas==3.0.1
python-dotenv==1.2.1
scikit_learn==1.8.0
seaborn==0.13.2
SQLAlchemy==2.0.46
tqdm==4.67.3
